completed with the milestone of Insert entity in React app
